+++ packages/client/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `client`: rename `watchBlockBlody` to `watchBlockBody`
 - `getEstimatedFee` takes as input a PolkadotSigner and optionally the "hinted-sign-extensions"
+- `tx`: now transactions are mortal by default with a 64 blocks period
 
 ### Changed